error in abhay laptop

+++ src/main/java/com/food/Entities/Cart.java
@@ -20,7 +20,7 @@ public class Cart
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer cartId;
-    private Integer cartTotal;
+    private Double cartTotal;
     private Integer cartSize;
     private Integer totalCartDiscount;
